step5 - refactor domain test package 로 이동
