+++ src/test/java/study/racing4/domain/CarRacingTest.java
@@ -1,4 +1,4 @@
-package study.racing4;
+package study.racing4.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
